pyproject.toml
# FILE: customer_support_rag/pyproject.toml

[tool.poetry]
name = "customer-support-rag"
version = "0.1.0"
description = "Production-Ready Multi-Agent RAG Customer Support System powered by LangGraph + LlamaIndex"
authors = ["AI Engineer <[truong311203@gmail.com]>"]
readme = "README.md"
packages = [
    { include = "config" },
    { include = "core" },
    { include = "ingestion" },
    { include = "retrieval" },
    { include = "agents" },
    { include = "api" },
]

[tool.poetry.dependencies]
python = "^3.11"

# ── LlamaIndex Core & Integrations ─────────────────────────────────────────
llama-index-core = ">=0.10"
llama-index-vector-stores-chroma = ">=0.1"
llama-index-embeddings-openai = ">=0.1"
llama-index-readers-file = ">=0.1"
# ── LangGraph & LangChain (Agent Orchestration) ────────────────────────────
langgraph = ">=0.2"
langchain-core = ">=0.2"
langchain-openai = ">=0.1"

# ── Vector Store ────────────────────────────────────────────────────────────
chromadb = ">=0.5"

# ── Reranking ───────────────────────────────────────────────────────────────
cohere = ">=5.5"                             # Cohere Rerank API
llama-index-postprocessor-cohere-rerank = ">=0.1"  # LlamaIndex Cohere postprocessor

# ── Sparse Retrieval ────────────────────────────────────────────────────────
rank-bm25 = ">=0.2"
# ── LLM Providers ──────────────────────────────────────────────────────────
openai = ">=1.30"
ollama = ">=0.2"                         # Future expansion — Ollama local LLM

# ── Data Validation & Config ────────────────────────────────────────────────
pydantic = ">=2.7"
pydantic-settings = ">=2.3"

# ── Structured Logging ─────────────────────────────────────────────────────
structlog = ">=24.0"

# ── API Layer ───────────────────────────────────────────────────────────────
fastapi = ">=0.111"
uvicorn = { extras = ["standard"], version = ">=0.30" }
httpx = ">=0.27"
python-multipart = ">=0.0.9"

# ── Simple UI ───────────────────────────────────────────────────────────────
streamlit = ">=1.30"

# ── Utilities ───────────────────────────────────────────────────────────────
tenacity = ">=8.3"
tiktoken = ">=0.7"
python-dotenv = ">=1.0"
aiofiles = ">=23.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2"
pytest-asyncio = ">=0.23"
pytest-cov = ">=5.0"
ruff = ">=0.4"
mypy = ">=1.10"
types-aiofiles = ">=23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=. --cov-report=term-missing --cov-report=html"

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
